Fix Device CRUD

+++ AquariumApi.DeviceApi/Services/HardwareService.cs
@@ -1,5 +1,6 @@
 ﻿using AquariumApi.Models;
 using Bifrost.IO.Ports;
+using Microsoft.AspNetCore.Hosting;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.Logging;
 using System;
@@ -18,12 +19,14 @@ namespace AquariumApi.DeviceApi
         private readonly IConfiguration _config;
         private readonly ILogger<HardwareService> _logger;
         private readonly ISerialService _serialService;
+        private readonly IHostingEnvironment _hostingEnvironment;
 
-        public HardwareService(IConfiguration config, ILogger<HardwareService> logger,ISerialService serialService)
+        public HardwareService(IConfiguration config, ILogger<HardwareService> logger,ISerialService serialService,IHostingEnvironment hostingEnvironment)
         {
             _config = config;
             _logger = logger;
             _serialService = serialService;
+            _hostingEnvironment = hostingEnvironment;
         }
         public AquariumDevice ScanHardware()
         {
@@ -43,7 +46,7 @@ namespace AquariumApi.DeviceApi
             try
             {
                 File.Delete("hardwaretest.jpg");
-                $"/usr/bin/raspistill -o hardwaretest.jpg".Bash();
+                $"/usr/bin/raspistill -w 100 -h 100 -o hardwaretest.jpg".Bash();
                 if(File.Exists("hardwaretest.jpg"))
                 {
                     File.Delete("hardwaretest.jpg");
@@ -59,6 +62,8 @@ namespace AquariumApi.DeviceApi
         public byte[] TakePhoto(CameraConfiguration config)
         {
             config.Output = "temp.jpg";
+            if (_hostingEnvironment.IsDevelopment())
+                return System.IO.File.ReadAllBytes(config.Output);
             //Path.GetDirectoryName(config.Output);
             if (File.Exists(config.Output))
                 File.Delete(config.Output);
